# docs/swagger.yaml
info:
  contact: {}
paths:
  /blog/addBlog:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: blog名称
        in: formData
        name: name
        type: string
      - description: blog标签
        in: formData
        name: label
        type: string
      - description: blog描述
        in: formData
        name: describe
        type: string
      - description: blog描述1
        in: formData
        name: describe1
        type: string
      - description: blog图片1
        in: formData
        name: image1
        type: file
      - description: blog描述2
        in: formData
        name: describe2
        type: string
      - description: blog图片2
        in: formData
        name: image2
        type: file
      - description: blog描述3
        in: formData
        name: describe3
        type: string
      - description: blog图片3
        in: formData
        name: image3
        type: file
      - description: blog描述4
        in: formData
        name: describe4
        type: string
      - description: blog图片4
        in: formData
        name: image4
        type: file
      - description: blog描述5
        in: formData
        name: describe5
        type: string
      - description: blog图片5
        in: formData
        name: image5
        type: file
      - description: blog描述6
        in: formData
        name: describe6
        type: string
      - description: blog图片6
        in: formData
        name: image6
        type: file
      - description: blog描述7
        in: formData
        name: describe7
        type: string
      - description: blog图片7
        in: formData
        name: image7
        type: file
      - description: blog描述8
        in: formData
        name: describe8
        type: string
      - description: blog图片8
        in: formData
        name: image8
        type: file
      - description: blog描述9
        in: formData
        name: describe9
        type: string
      - description: blog图片9
        in: formData
        name: image9
        type: file
      - description: blog描述10
        in: formData
        name: describe10
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加新的blog
      tags:
      - blog模块
  /blog/deleteBlog:
    post:
      parameters:
      - description: blogID
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除blog
      tags:
      - blog模块
...
  /blog/getTheBlog:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: blogID
        in: formData
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 获取指定的blog
      tags:
      - blog模块
  /blog/updateBlog:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: blogID
        in: formData
        name: id
        type: string
      - description: blog名称
        in: formData
        name: name
        type: string
      - description: blog标签
        in: formData
        name: label
        type: string
      - description: blog描述
        in: formData
        name: describe
        type: string
      - description: blog描述1
        in: formData
        name: describe1
        type: string
      - description: blog图片1
        in: formData
        name: image1
        type: file
      - description: blog描述2
        in: formData
        name: describe2
        type: string
      - description: blog图片2
        in: formData
        name: image2
        type: file
      - description: blog描述3
        in: formData
        name: describe3
        type: string
      - description: blog图片3
        in: formData
        name: image3
        type: file
      - description: blog描述4
        in: formData
        name: describe4
        type: string
      - description: blog图片4
        in: formData
        name: image4
        type: file
      - description: blog描述5
        in: formData
        name: describe5
        type: string
      - description: blog图片5
        in: formData
        name: image5
        type: file
      - description: blog描述6
        in: formData
        name: describe6
        type: string
      - description: blog图片6
        in: formData
        name: image6
        type: file
      - description: blog描述7
        in: formData
        name: describe7
        type: string
      - description: blog图片7
        in: formData
        name: image7
        type: file
      - description: blog描述8
        in: formData
        name: describe8
        type: string
      - description: blog图片8
        in: formData
        name: image8
        type: file
      - description: blog描述9
        in: formData
        name: describe9
        type: string
      - description: blog图片9
        in: formData
        name: image9
        type: file
      - description: blog描述10
        in: formData
        name: describe10
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改指定的blog
      tags:
      - blog模块
  /experience/addExperience:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 经验名字
        in: formData
        name: name
        type: string
      - description: 经验网址
        in: formData
        name: url
        type: string
      - description: 经验描述
        in: formData
        name: description
        type: string
      - description: 经验时间
        in: formData
        name: time
        type: string
      - description: 经验图片
        in: formData
        name: image
        type: file
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加新的经验
      tags:
      - 经验模块
...
  /message/addMessage:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 名字
        in: formData
        name: name
        type: string
      - description: 留言内容
        in: formData
        name: text
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加新的留言
      tags:
      - 留言模块
...
swagger: "2.0"